Thanks to user attractivechaos, I was able to remove a pointer in HashTableNode. It is literally as memory efficient as possible now.
The HashTableNode structure now only has one pointer making it as memory efficient as possible. The only caveat to make this work is I had to remove the function: "hashtable_remove". This function did provide convenience in some situations over its counterpart "hashtable_remove_key", but the two are interchangeable EVERYWHERE, so I went ahead and made this change since the pros outweighed the cons.